Add comment, update VisualPanel onItemClick to have index value for callback function. Move VisualItem into components from VisualPanel folder. Make TestImages for VisualPanel together in one group
